Update patch set 5
Patch Set 5:

(7 comments)

> not sure how to force asterisk binary version here
 > 
 > We have:
 > asterisk-dev (= 1:11.13.1~dfsg-2) [all]
 > asterisk (= 1:11.13.1~dfsg-2+b1) [amd64]
 > 
 > See previous revisions in order to see my failures

Depends what you want to achieve? Do you want to have exact versions as the ones currently in the archive? Or be more lax and allow for some version variance?

The problem you've seen with the substvar is due to missing asterisk on the build environment. The fix should be simply to build-depend on asterisk, which is a bit heavy weight but oh well, and restore the substvar generation in debian/rules.

Patch-set: 5