Update patch set 2
Patch Set 2: Code-Review+2

Yeah, an equal version is very strict, and because we are only going to see a single version in the archive, restricting just makes it more difficult to try to fix the code (if there was a need) with newer asterisk versions. I think we should ignore the build failures here and mark as verified.

Patch-set: 2
Reviewer: Guillem Jover <1000016@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=+2