Update patch set 6
Patch Set 6:

this review builds correctly on stretch, the only thing that prevents jenkins to build it is "upstream" which contains the different version of app_voicemail.c, but as soon as we change it there we break the current trunk version  I think. Need your experience how to solve that.

Patch-set: 6