Update patch set 6
Patch Set 6:

> > this review builds correctly on stretch,

Thanks for working on this, Kirill!

 > > the only thing that
 > > prevents jenkins to build it is "upstream" which contains the
 > > different version of app_voicemail.c,

Hm, the apps/app_voicemail.c as present in this review, is this from you or copied/derived from somewhere else?

 > > but as soon as we change it
 > > there we break the current trunk version  I think. Need your
 > > experience how to solve that.
 
If we'd patch apps/app_voicemail.c from upstream using debian/patches/... - would this basically work, nor?
But it would fail then to compile/work against our jessie (based) versions?

 > Perhaps we should have a repo per major asterisk version? Then we
 > can have both coexist, and don't need to bother with conditional
 > patching or similar ugliness? So we'd have something like
 > asterisk-11-voicemail and then asterisk-13-voicemail (or a naming
 > on those lines).

I tend to think that we should avoid this if possible, otherwise we'll end up with multiple repositories which we don't really care about or have a mess with maintaining (I'm thinking of the sems situation we currently have). :-/

Patch-set: 6