TT#3660 Switch the package to be a simple overlay over upstream
Move to a development mode where we do not carry any pristine upstream
code, nor depend on any specific version. We instead carry patch series
against major versions of asterisk. And we use 0 fuzz so that we can
catch potentially broken combinations of our code aginst the upstream
baseline.

This way we can build against multiple upstream major versions, and in
case some are compatible we can simply use symlinks for the patches
directory.

Bump the version because we are changing the "upstream" part of the
repository.

Change-Id: I5015883416ba0f94db66f5d5123f61caa09f22a1