Update patch set 2
Patch Set 2:

AFAICT the build flags must be different. If both ODBC_STORAGE and IMAP_STORAGE are set, then STORE() would point to the ODBC store() function, but `chan` would still be defined. If IMAP_STORAGE is not set but only ODBC_STORAGE is set, then `chan` would be missing but still the ODBC store() function would be used (which results in error as it also needs `chan`).

Patch-set: 2
Label: Verified=0