MT#60926 Cleanup patch series (part 1)
This rebases the patch series, reworks them so that fixes for
previous changes get integrated into the initial change, and renames
patches to have uniform names. This should make the patches easier to
understand.

This is just an initial pass, further cleanup might be done in the
future.

The easiest way to see the actual differences in the end result is
to run «gbp pq import» into a branch, before and after this patch
series, and «git diff» both branches.

Change-Id: I5d1cace348f2195a8d8f376c9e7ebb03e3d5eace